Feat: 캐러셀 모듈화 및 에러 수정
